Removes convmatrix and loadstats modules + adds mix tests
